Moved components into more specifiy sub-directories/namespaces. Currently divided into "GamePage" and "General".
