+++ src/Codebreaker.WinUI/Views/Components/GamePage/GameResultDisplay.xaml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <UserControl
-    x:Class="CodeBreaker.WinUI.Views.Components.GameResultDisplay"
+    x:Class="CodeBreaker.WinUI.Views.Components.GamePage.GameResultDisplay"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
