+++ src/Codebreaker.WinUI/Views/Components/GamePage/PegSelectionComponent.xaml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <UserControl
-    x:Class="CodeBreaker.WinUI.Views.Components.PegSelectionComponent"
+    x:Class="CodeBreaker.WinUI.Views.Components.GamePage.PegSelectionComponent"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:local="using:CodeBreaker.WinUI.Views.Components"
