+++ src/Codebreaker.WinUI/Views/Components/General/InfoBarArea.xaml
@@ -1,9 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <UserControl
-    x:Class="CodeBreaker.WinUI.Views.Components.InfoBarArea"
+    x:Class="CodeBreaker.WinUI.Views.Components.General.InfoBarArea"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="using:CodeBreaker.WinUI.Views.Components"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     xmlns:vmComponents="using:Codebreaker.ViewModels.Components"
