+++ src/Codebreaker.WinUI/Views/Pages/ShellPage.xaml
@@ -7,7 +7,7 @@
     xmlns:helpers="using:CodeBreaker.WinUI.Helpers"
     xmlns:behaviors="using:CodeBreaker.WinUI.Behaviors"
     xmlns:i="using:Microsoft.Xaml.Interactivity"
-    xmlns:components="using:CodeBreaker.WinUI.Views.Components"
+    xmlns:generalComponents="using:CodeBreaker.WinUI.Views.Components.General"
     xmlns:toolkitConverters="using:CommunityToolkit.WinUI.UI.Converters"
     xmlns:cme="using:CodeBreaker.WinUI.CustomMarkupExtensions"
     xmlns:ca="using:CodeBreaker.WinUI.CustomAttachedProperties"
@@ -78,7 +78,7 @@
                     <RowDefinition Height="Auto" />
                     <RowDefinition />
                 </Grid.RowDefinitions>
-                <components:InfoBarArea />
+                <generalComponents:InfoBarArea />
                 <Frame Grid.Row="1" x:Name="NavigationFrame" />
             </Grid>
             <NavigationView.FooterMenuItems>
